+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file, and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.0 - 2022-01-12
+
+* Added the `RequireImport` and `ForbidImport` annotations per [#123](https://github.com/microsoft/pybryt/issues/123)
+* Added `TimeComplexityChecker` for analyzing complexity without annotations per [#137](https://github.com/microsoft/pybryt/issues/137)
+* Fixed annotation message filtering bug per [#145](https://github.com/microsoft/pybryt/issues/145)
+* Added `dataclasses` backport to required modules for Python &lt; 3.7
+* Added a `group` argument to the `check` context manager per [#146](https://github.com/microsoft/pybryt/issues/146)
+* Moved named annotation filtering into `ReferenceImplementation` constructor per [#147](https://github.com/microsoft/pybryt/issues/147)
+
 ## 0.3.1 - 2021-12-01
 
 * Fixed a bug that was causing floats to be added to the memory footprint prematurely per resulting in [#125](https://github.com/microsoft/pybryt/issues/125)